updating integration tests to deal with templating in plotNames
