+++ test/pythonml/test_post_processors.py
@@ -8,28 +8,6 @@ from parameterized import parameterized
 from unittest_baseclass import BaseUnitTest
 
 
-def template_plot_names(flavor, plot_names):
-    """
-    This function templates in plot_names into flavor file. This is needed because the
-    names of the plots are now set by users via a FileContentResults Form
-
-    Args:
-        flavor (str): name of the python model flavor file in assets
-        plot_names (str): name of the plot that is made by the flavor, and it is for
-            this plot we check
-    """
-
-    assert (os.path.isfile(flavor))
-    with open(flavor, "r") as input:
-        raw_template_filelines = input.readlines()
-
-    with open(flavor, "w") as output:
-        for line in raw_template_filelines:
-            for index, plot_name in enumerate(plot_names):
-                line = re.sub("{{ results\[*"+str(index)+"].basename }}", plot_name, line)
-            output.write(line)
-
-
 class TestPostProcessingFlavors(BaseUnitTest):
     """
     This class performs unit tests for the post-processing flavors
@@ -48,7 +26,7 @@ class TestPostProcessingFlavors(BaseUnitTest):
         """
         self.set_pickle_fixtures_path_in_context_object(category, 'model_data')
         shutil.copy(os.path.join(self.asset_path, flavor), flavor)
-        template_plot_names(flavor, plots)
+        BaseUnitTest.template_plot_names(flavor, plots)
         subprocess.call(['python', flavor])
         for plot in plots:
             assert os.path.isfile(plot)
