+++ test/pythonml/unittest_baseclass.py
@@ -99,6 +99,29 @@ class BaseUnitTest(unittest.TestCase):
             names = ['train_descriptors', 'test_descriptors', 'train_target', 'test_target', 'descriptors', 'label_encoder']
         return names
 
+    @staticmethod
+    def template_plot_names(flavor, plot_names):
+        """
+        This function templates in plot_names into flavor file. This is needed because the
+        names of the plots are now set by users via a FileContentResults Form.
+
+        Args:
+            flavor (str): name of the python model flavor file in assets.
+            plot_names (str): name of the plot that is made by the flavor, and it is for
+                this plot we check.
+        """
+
+        assert (os.path.isfile(flavor))
+        with open(flavor, "r") as input:
+            raw_template_filelines = input.readlines()
+
+        with open(flavor, "w") as output:
+            for line in raw_template_filelines:
+                for index, plot_name in enumerate(plot_names):
+                    line = re.sub("{{ results\[*" + str(index) + "].basename }}", plot_name, line)
+                output.write(line)
+
+
     def set_pickle_fixtures_path_in_context_object(self, category, data_type):
         """
         This function updated the paths in the context object with  the 'names' of
